Migrate Lineage functionality test to ProgramMetadataTest
